pyproject.toml
[tool.poetry]
name = "blockpi-proxy"
version = "0.0.1"
description = "Simple reverse proxy for BlockPi endpoints"
authors = ["smk762 <smk@komodoplatform.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
uvicorn = "^0.29.0"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
httpx = "^0.27.0"
requests = "^2.31.1"
pytest-env = "^1.1.3"
fastapi-utils = "0.6.0"
pytest = "^8.2.1"


[tool.poetry.group.dev.dependencies]
black = "^24.4.2"
pytest = "^8.2.1"
pytest-cov = "^5.0.0"
flake8 = "^7.0.0"
pytest-flake8 = "^1.1.1"



[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
